Have issue reporter use service injection
The main part of this change is pulling out the IssueReporter class into its own file and then having the Main.ts file be all about initialization.

This will be easier to register remote services (to help get rid of all the raw IPC)
